Split rest of Monitor classes into Interface and an implementation
